split relationships
